+++ libpthread/pthread_sys_write.c
@@ -3,7 +3,7 @@
 #include <pthread.h>
 #include "thread_internal.h"
 
-int write(int fd, const void *buf, size_t count) {
+ssize_t write(int fd, const void *buf, size_t count) {
   __TEST_CANCEL();
   return __libc_write(fd,buf,count);
 }
